IO

Input

KitAMR.read_config — Function

read_config(filename::AbstractString) -> Dict{Symbol, Any}

Read .txt configuration file and return a dictionary.

KitAMR.listen_for_save! — Function

listen_for_save!(; rank)

Start listening for the input from command line. Must be called before calling check_for_save!.

KitAMR.check_for_save! — Function

check_for_save!(ps4est, amr; rank)

Check whether save is input during the iteration. If the result is true, a saving process will be executed.

Todo

A bug exists here, which causes the save input will not be aware of, until one stdout is invoked.

Output

KitAMR.save_result — Function

save_result(
    ps4est::Ptr{KitAMR.P4est.LibP4est.p4est},
    amr::KitAMR_Data{DIM, NDF};
    dir_path
)

Save all results to dir_path. If isempty(dir_path) is true, a folder named result<yyyy-mm-dd_HH-MM> will be made in current path.