Data

All of the data in KitAMR.jl are organized as follows:

KitAMR.KitAMR_Data — Type

mutable struct KitAMR_Data{DIM, NDF}

Structure that collects all of the data.

global_data::Global_Data

ghost::Ghost

field::Field

KitAMR.Global_Data — Type

mutable struct Global_Data{DIM, NDF}

Structure of required information during a simulation.

config::Configure: Defined by Configure.

forest::Forest: Defined by Forest.

status::Status: Defined by Status.

KitAMR.Forest — Type

mutable struct Forest{DIM}

Structure related to p4est.

p4est::Union{Ptr{KitAMR.P4est.LibP4est.p4est}, Ptr{KitAMR.P4est.LibP4est.p8est}}

ghost::Union{Ptr{p4est_ghost_t}, Ptr{p8est_ghost_t}}

mesh::Union{Ptr{p4est_mesh_t}, Ptr{p8est_mesh_t}}

KitAMR.Status — Type

mutable struct Status

Structure of simulation status. It contains the real time information, and is updated every step.

Fields

max_vs_num::Int64: Maximum number of the velocity grids among ghost quadrants.

gradmax::Vector{Float64}: Maximum absolute value of the gradients of conserved variables.

wmax::Vector{Float64}: Maximum value of conserved variables.

wmin::Vector{Float64}: Minimum value of conserved variables.

Δt::Float64: Time step size used in iterate!.

Δt_ξ::Float64: Time step size constraint by grid size.

sim_time::Float64: Dimensionless simulation time.

ps_adapt_step::Int64: Number of steps after last AMR in physical space.

vs_adapt_step::Int64: Number of steps after last AMR in velocity space.

partition_step::Int64: Number of steps after last partition.

residual::Residual: Residual of conserved variables defined by Residual.

save_flag::Base.RefValue{Bool}: Flag indicating whether to save.

KitAMR.Residual — Type

mutable struct Residual

step::Int64: Current step number.

residual::Vector{Float64}: Residual of primary variables, is updated by residual_check!.

sumRes::Vector{Float64}: Summation of residuals.

sumAvg::Vector{Float64}: Summation of primary variables to average sumRes.

redundant_step::Int64: Count of redundant step to insure the convergence.

KitAMR.Ghost — Type

mutable struct Ghost

Structure of ghost layer for communication between processors.

ghost_exchange::Ghost_Exchange

ghost_wrap::Vector{KitAMR.AbstractGhostPsData}

KitAMR.Ghost_Exchange — Type

mutable struct Ghost_Exchange

Structure of ghost data and pointers used by p4est.

ghost_datas::Any

ghost_slopes::Any

ghost_structures::Any

mirror_data_pointers::Any

mirror_slope_pointers::Any

mirror_structure_pointers::Any

KitAMR.Field — Type

mutable struct Field{DIM, NDF}

Structure of fields data.

trees::PS_Trees: Cells data defined by PS_Trees.

faces::Vector{AbstractFace}: Mapping between faces and cells. The element type is defined by AbstractFace.

immersed_boundary::KitAMR.ImmersedBoundary

KitAMR.PS_Trees — Type

mutable struct PS_Trees{DIM, NDF}

Structure of cells data managed by Julia.

data::Array{Array{AbstractPsData{DIM, NDF}, 1}, 1} where {DIM, NDF}

offset::Int64: Offset of treeid used in partition.