Initialization

KitAMR.initialize_KitAMR — Function

initialize_KitAMR(
    config::Dict
) -> Tuple{Union{Ptr{KitAMR.P4est.LibP4est.p4est}, Ptr{KitAMR.P4est.LibP4est.p8est}}, KitAMR_Data}

Initialize everthing according to config dictionary.

KitAMR.initialize_ghost — Function

initialize_ghost(
    p4est::Union{Ptr{KitAMR.P4est.LibP4est.p4est}, Ptr{KitAMR.P4est.LibP4est.p8est}},
    global_data::Global_Data
) -> Ghost

Initialize Ghost structure.

KitAMR.initialize_field! — Function

initialize_field!(
    global_data::Global_Data{DIM, NDF}
) -> Tuple{Any, Union{Ptr{KitAMR.P4est.LibP4est.p4est}, Ptr{KitAMR.P4est.LibP4est.p8est}}}

Initialize field for 2D case.

initialize_field!(
    global_data::Global_Data{3, NDF}
) -> Tuple{Any, Union{Ptr{KitAMR.P4est.LibP4est.p4est}, Ptr{KitAMR.P4est.LibP4est.p8est}}}

Initialize field for 3D case.

In this function, p4est is initialized. Mesh (in both physical and velocity space) is generated and adaptively refined. Field (in both physical and velocity space) is initialized according to initial conditions. Mapping between Julia data and p4est is established. Currently, realizations for 2D and 3D cases are different. The method for 3D case can be more complicated but more efficient. The method for 2D will be improved to be consistent with the 3D case.

Initialize PS_Data and establish the mapping between Julia data and p4est.

KitAMR.initialize_ps! — Function

initialize_ps!(
    ps4est::Ptr{KitAMR.P4est.LibP4est.p4est},
    global_data::Global_Data{DIM, NDF}
) -> PS_Trees

initialize_ps!(
    p4est::Ptr{KitAMR.P4est.LibP4est.p8est},
    global_data::Global_Data{DIM, NDF}
) -> Any

Mesh identification, geometric adaptive refinement, and initial partition.

KitAMR.pre_refine! — Function

pre_refine!(
    ps4est::Ptr{KitAMR.P4est.LibP4est.p4est},
    global_data::Global_Data
)

pre_refine!(
    ps4est::Ptr{KitAMR.P4est.LibP4est.p8est},
    global_data::Global_Data
) -> Any

Initialize faces defined by AbstractFace.

KitAMR.initialize_faces! — Function

initialize_faces!(
    ps4est::Ptr{KitAMR.P4est.LibP4est.p4est},
    amr::KitAMR_Data
)

initialize_faces!(
    ps4est::Ptr{KitAMR.P4est.LibP4est.p8est},
    amr::KitAMR_Data
)

Initialize the ghost layer in p4est.

KitAMR.AMR_ghost_new — Function

AMR_ghost_new(
    p4est::Ptr{KitAMR.P4est.LibP4est.p4est}
) -> Ptr{p4est_ghost_t}

AMR_ghost_new(
    p4est::Ptr{KitAMR.P4est.LibP4est.p8est}
) -> Ptr{p8est_ghost_t}

Initialize the mesh in p4est that is used to look up the neighboring cell in physical space.

KitAMR.AMR_mesh_new — Function

AMR_mesh_new(
    p4est::Ptr{KitAMR.P4est.LibP4est.p4est},
    ghost::Ptr{p4est_ghost_t}
) -> Ptr{p4est_mesh_t}

AMR_mesh_new(
    p4est::Ptr{KitAMR.P4est.LibP4est.p8est},
    ghost::Ptr{p8est_ghost_t}
) -> Ptr{p8est_mesh_t}

Initialize Neighbor in PS_Data.

KitAMR.initialize_neighbor_data! — Function

initialize_neighbor_data!(
    ip::PointerWrapper{p4est_iter_volume_info},
    data,
    dp
)

initialize_neighbor_data!(
    ip::PointerWrapper{p8est_iter_volume_info},
    data,
    dp
)

initialize_neighbor_data!(info, data)

initialize_neighbor_data!(
    ps4est::Union{Ptr{KitAMR.P4est.LibP4est.p4est}, Ptr{KitAMR.P4est.LibP4est.p8est}},
    amr::KitAMR_Data
)

Initialize SolidNeighbors corresponding to the immersed boundaries.

KitAMR.initialize_solid_neighbor! — Function

initialize_solid_neighbor!(amr::KitAMR_Data)

initialize_solid_neighbor!(
    ps_data::PS_Data{DIM, NDF},
    amr::KitAMR_Data{DIM, NDF}
)