Theory

The backend functions related to kinetic theory are integrated in lib/KitCore. Some interfaces are provided

KitAMR.get_prim — Function

get_prim(
    ps_data::PsData{2, NDF},
    kinfo::KInfo{2, NDF}
) -> Vector{Float64}

Get primary macroscopic variables from the conserved variables.

KitAMR.get_conserved — Function

get_conserved(
    ps_data::PsData{2, NDF},
    kinfo::KInfo
) -> Vector{Float64}

Get conserved macroscopic variables from the primary variables.

KitAMR.calc_w0 — Function

calc_w0(
    midpoint::AbstractMatrix,
    df::AbstractMatrix,
    weight::AbstractVector,
    _::KInfo{2, 2}
) -> Vector

Calculate the conserved macroscopic variables according to the microscopic distribution.

KitAMR.calc_qf — Function

calc_qf(
    vs_data::AbstractVsData{2, 2},
    prim::AbstractVector
) -> Vector

Calculate the heat flux according to the microscopic distribution.

KitAMR.discrete_maxwell — Function

discrete_maxwell(
    ps_data::PsData,
    kinfo::KInfo
) -> Matrix{Float64}

Caculate the discretized Maxwellian distribution according to the primary macroscopic variables.

KitAMR.shakhov_part — Function

shakhov_part(
    midpoint::AbstractMatrix,
    F::AbstractMatrix,
    prim::AbstractVector,
    qf::AbstractVector,
    kinfo::KInfo{2, 2}
) -> Union{Matrix, Vector{Float64}}

Calculate the discretized Shakhov correction.

KitAMR.ref_vhs_vis — Function

ref_vhs_vis(Kn, omega) -> Any
ref_vhs_vis(Kn, omega, alpha) -> Any
ref_vhs_vis(Kn, omega, alpha, T_ref) -> Any

Reference viscousity obtained by VHS molecules model.

KitAMR.get_τ — Function

get_τ(prim::AbstractVector, μ::Real, ω::Real) -> Any

Calculate collisional time.