Output configuration

KitAMR.jl outputs the field as .jld2 and .pvtu files for post-processing.

KitAMR.Output — Type

mutable struct Output

Structure of output information.

Fields

vtk_celltype::Type{Tp} where Tp<:KitAMR.AbstractVTKCellType: VTK cell type in physical space. Available options: Pixel and Triangle for 2D; Voxel and Tetra for 3D. Default is Triangle for 2D and Tetra for 3D.

vs_vtk_celltype::Type{Tv} where Tv<:KitAMR.AbstractVTKCellType: VTK cell type in velocity space. Available options: Pixel for 2D.

anim_dt::Float64: Time interval between animation frames.

vs_output_criterion::Function: User-defined-function determining the physical cells need to output the velocity space.

anim_index::Int64: Index of the last saved animation frame.

Supported vtk cell types are

KitAMR.Pixel — Type

abstract type Pixel <: KitAMR.AbstractVTKCellType

KitAMR.Voxel — Type

abstract type Voxel <: KitAMR.AbstractVTKCellType

KitAMR.Triangle — Type

abstract type Triangle <: KitAMR.AbstractVTKCellType

KitAMR.Tetra — Type

abstract type Tetra <: KitAMR.AbstractVTKCellType

Note that the vertices in .pvtu file are redundant. To obtain contour of the field, a cleaning process in post-process software is required. For example, using Paraview, one needs to execute Clean to Grid before contour.

Triangle and Tetra are chosen for smoother field visualization, but will not improve the accuracy of the solution. The variables on their vertices are linearly reconstructed.