Configuration of problem

To simulate a problem using KitAMR, user is required to define the problem. Most of the user-provided information is collected in Configure struct.

KitAMR.Configure — Type

struct Configure{DIM, NDF}

Structure of problem configuration. DIM and NDF represent the dimension number of the problem and the number of the distribution function. This struct plays an key role in the solution process.

Fields

geometry::Vector{Float64}: The range of the simulated domain. As an example, for 2D case, it should be aligned as [xmin,xmax,ymin,ymax].

trees_num::Vector{Int64}: The number of tree roots for each dimension. For 2D case, it should be aligned as [xnum,ynum]

quadrature::Union{KitAMR.AbstractQuadrature, Vector{Float64}}: For Vector type, it represents the range of the velocity space. For 2D case, it should be aligned as [umin,umax,vmin,vmax]. The AbstractQuadrature type is reserved for other quadrature rule like Gauss-Hermite.

vs_trees_num::Vector{Int64}: The number of tree roots for each dimension in velocity space.

IC::KitAMR.AbstractInitCondType: The initial` condition.

domain::Vector{KitAMR.Domain}: The types of the domain boundary. For 2D case, the vector should catain 4 elements corresponding to the 4 domain boundaries.

IB::Vector{KitAMR.AbstractBoundary}: The immersed boundary. Multiple boundaries are supported.

gas::KitAMR.Gas: The property of the simulated gas.

solver::KitAMR.Solver: The set of the solver.

output::KitAMR.Output: The set of the output form.

user_defined::KitAMR.UDF: The functions defined by users, including some criteria.