Configuration

To simulate a problem using KitAMR, user is required to define the problem. Most of the user-provided information is collected in Configure struct.

KitAMR.Configure — Type

struct Configure{DIM, NDF}

Structure of problem configuration. DIM and NDF represent the dimension number of the problem and the number of the distribution function. This struct plays an key role in the solution process.

Fields

geometry::Vector{Float64}: Range of the simulated domain. As an example, for 2D case, it should be aligned as [xmin,xmax,ymin,ymax].

trees_num::Vector{Int64}: Number of tree roots for each dimension. For 2D case, it should be aligned as [xnum,ynum]

quadrature::Union{KitAMR.AbstractQuadrature, Vector{Float64}}: For Vector type, it represents the range of the velocity space. For 2D case, it should be aligned as [umin,umax,vmin,vmax]. The AbstractQuadrature type is reserved for other quadrature rule like Gauss-Hermite.

vs_trees_num::Vector{Int64}: Number of tree roots for each dimension in velocity space.

IC::AbstractInitCondType: Initial condition defined by AbstractInitCondType.

domain::Vector{Domain}: Types of the domain boundary. For 2D case, the vector should catain 4 elements corresponding to the 4 domain boundaries. The element type is defined by Domain.

IB::Vector{AbstractBoundaryType}: Immersed boundary. Multiple boundaries are supported. The element type is defined by AbstractBoundaryType

gas::Gas: Properties of the simulated gas defined by Gas.

solver::Solver: Setup of the solver defined by Solver.

output::Output: Setup of the output form defined by Output.

user_defined::KitAMR.UDF: Functions defined by users, including some criteria.

KitAMR.Solver — Type

struct Solver

Structure of solver configuration.

Fields

CFL::Float64: Courant-Friedrichs-Lewy number.

AMR_PS_MAXLEVEL::Int64: Maximum level of the (static) AMR in physical space.

AMR_DYNAMIC_PS_MAXLEVEL::Int64: Maximum level of the dynamic AMR in physical space. In most cases, it should be smaller than the static one. The default is equal to AMR_PS_MAXLEVEL.

AMR_VS_MAXLEVEL::Int64: Maximum level of the AMR in velocity space.

flux::KitAMR.AbstractFluxType: The numerical flux type.

time_marching::KitAMR.AbstractTimeMarchingType: The time-marching scheme.

PS_DYNAMIC_AMR::Bool: The dynamic AMR in physical space is open or not. The default is true.

VS_DYNAMIC_AMR::Bool: The dynamic AMR in velocity space is open or not. The default is true.